+++ CHANGELOG.md
@@ -22,6 +22,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Fixed ESLint warnings:
   * `core/component/render-function`
   * `core/component/register`
+  * `core/component/directives/update-on`
 
 ## v3.0.0-rc.47 (2020-07-31)